norma i ordre: debugs a banda
