+++ src/builtins/env.c
@@ -6,7 +6,7 @@
 /*   By: ecoma-ba <ecoma-ba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/04 15:37:26 by ecoma-ba          #+#    #+#             */
-/*   Updated: 2025/01/04 15:37:27 by ecoma-ba         ###   ########.fr       */
+/*   Updated: 2025/01/04 15:46:53 by ecoma-ba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
